backend/requirements.txt
Flask
Flask-cors
flask-sqlalchemy